Refactor: getFormattedTime helper.js로 분리
