+++ src/components/utils/helper.jsx
@@ -5,3 +5,11 @@ export const getFormattedDate = () => {
   const day = date.getDate();
   return `${year}.${month}.${day}`;
 };
+
+export const getFormattedTime = (time) => {
+  const minutes = Math.floor(time / 60);
+  const seconds = time % 60;
+  return `${minutes < 10 ? '0' : ''}${minutes}:${
+    seconds < 10 ? '0' : ''
+  }${seconds}`;
+};
